types [nfc]: Move Outbox to new file.
