+++ src/types.js
@@ -3,10 +3,12 @@ import type { IntlShape } from 'react-intl';
 import type { DangerouslyImpreciseStyleProp } from 'react-native/Libraries/StyleSheet/StyleSheet';
 
 import type { Auth, Topic, Message, Reaction, ReactionType, Narrow } from './api/apiTypes';
+import type { Outbox } from './outbox/outboxTypes';
 import type { AppStyles } from './styles/theme';
 
 export type * from './reduxTypes';
 export type * from './api/apiTypes';
+export type * from './outbox/outboxTypes';
 
 /**
  * Assert a contradiction, statically.  Do nothing at runtime.
@@ -149,94 +151,6 @@ export type TopicExtended = {|
 export type Context = {|
   styles: AppStyles,
 |};
-
-/**
- * A message we're in the process of sending.
- *
- * We use these objects for two purposes:
- *
- * (a) They make up the queue of messages the user has asked us to send, and
- *     which we'll retry sending if initial attempts fail.  See
- *     `trySendMessages`.
- *
- * (b) We show them immediately in the message list, even before we've
- *     successfully gotten them to the server (but with an activity
- *     indicator to show we're still working on them.)
- *
- * Even after (a) is complete for a given message, we still need the
- * `Outbox` object for the sake of (b), until we hear an `EVENT_NEW_MESSAGE`
- * event from the server that lets us replace it with the corresponding
- * `Message` object.
- *
- * This type most often appears in the union `Message | Outbox`, and so its
- * properties are deliberately similar to those of `Message`.
- */
-export type Outbox = {|
-  /** Used for distinguishing from a `Message` object. */
-  isOutbox: true,
-
-  /**
-   * False until we successfully send the message, then true.
-   *
-   * As described in the type's jsdoc (above), once we've sent the message
-   * we still keep the `Outbox` object around for a (usually short) time
-   * until we can replace it with a `Message` object.
-   */
-  isSent: boolean,
-
-  // These fields don't exist in `Message`.
-  // They're used for sending the message to the server.
-  markdownContent: string,
-  narrow: Narrow,
-
-  // These fields are modeled on `Message`.
-  avatar_url: string | null,
-  content: string,
-  display_recipient: $FlowFixMe, // `string` for type stream, else PmRecipientUser[].
-  id: number,
-  reactions: Reaction[],
-  sender_email: string,
-  sender_full_name: string,
-  subject: string,
-  timestamp: number,
-  type: 'stream' | 'private',
-|};
-
-/**
- * MessageLike: Imprecise alternative to `Message | Outbox`.
- *
- * Flow reasonably dispermits certain classes of access on union types. In
- * particular,
- * ```
- * const { sender_id } = (message: Message | Outbox);  // error
- * ```
- * is not allowed. However, as long as you're prepared to handle values of
- * `undefined`, it's both JavaScript-legal to do so and occasionally convenient.
- *
- * We therefore construct an intermediate type which Flow recognizes as a
- * subtype of `Message | Outbox`, but which Flow will permit us to directly (and
- * soundly) destructure certain `Message`-only fields from:
- * ```
- * const { sender_id } = (message: MessageLike);  // ok!
- * ```
- *
- * * Note: `MessageLike` <: `Message | Outbox`, but the converse does not hold.
- *   It is therefore strongly advised _never_ to use `MessageLike` as either an
- *   argument or return type; instead, always accept and produce values of
- *   `Message | Outbox`, and cast them to `MessageLike` at their use-site when
- *   necessary.
- *
- * * Note 2: This class is asymmetric mostly because there is no current use case
- *   for accessing Outbox-only fields on a `Message | Outbox`.
- *
- */
-export type MessageLike =
-  | $ReadOnly<Message>
-  | $ReadOnly<{
-      // $Shape<T> is unsound, per Flow docs, but $ReadOnly<$Shape<T>> is not
-      ...$Shape<{ [$Keys<Message>]: void }>,
-      ...Outbox,
-    }>;
 
 export type LocalizableText = string | { text: string, values?: { [string]: string } };
